+++ addons/rating/static/tests/helpers/mock_server/models/mail_message.js
@@ -1,11 +1,11 @@
 /** @odoo-module **/
 
-import '@mail/../tests/helpers/mock_server'; // ensure mail overrides are applied first
+import '@mail/../tests/helpers/mock_server/models/mail_message'; // ensure mail overrides are applied first
 
 import { patch } from "@web/core/utils/patch";
 import { MockServer } from "@web/../tests/helpers/mock_server";
 
-patch(MockServer.prototype, 'rating', {
+patch(MockServer.prototype, 'rating/models/mail_message', {
     /**
      * @override
      */
